# docs/swagger.yaml
basePath: /
definitions:
...
  protocol.ArticleVersion:
    properties:
      articleID:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      updatedAt:
        type: string
      version:
        type: integer
      versionID:
        type: integer
    type: object
...
  protocol.GetArticleVersionInfoResponse:
    properties:
      version:
        $ref: '#/definitions/protocol.ArticleVersion'
    type: object
...
  protocol.GetLatestArticleVersionInfoResponse:
    properties:
      version:
        $ref: '#/definitions/protocol.ArticleVersion'
    type: object
...
  protocol.HTTPResponse:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  protocol.ListArticleVersionsResponse:
    properties:
      pageInfo:
        $ref: '#/definitions/protocol.PageInfo'
      versions:
        items:
          $ref: '#/definitions/protocol.ArticleVersion'
        type: array
    type: object
...
  protocol.PageInfo:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
host: 9.134.115.68:8170
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Aris-blog API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Aris-blog
  version: "1.0"
paths:
...
  /v1/article/{articleID}/version/latest:
    get:
      consumes:
      - application/json
      parameters:
      - in: path
        name: articleID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取最新文章版本信息响应
          schema:
            $ref: '#/definitions/protocol.GetLatestArticleVersionInfoResponse'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取最新文章版本信息
      tags:
      - articleVersion
  /v1/article/{articleID}/version/list:
    get:
      consumes:
      - application/json
      parameters:
      - in: path
        name: articleID
        required: true
        type: integer
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 50
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 列出文章版本响应
          schema:
            $ref: '#/definitions/protocol.ListArticleVersionsResponse'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 列出文章版本
      tags:
      - articleVersion
  /v1/article/{articleID}/version/v{version}:
    get:
      consumes:
      - application/json
      parameters:
      - in: path
        name: articleID
        required: true
        type: integer
      - in: path
        minimum: 1
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取文章版本信息响应
          schema:
            $ref: '#/definitions/protocol.GetArticleVersionInfoResponse'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/protocol.HTTPResponse'
            - properties:
                data:
                  type: object
                error:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取文章版本信息
      tags:
      - articleVersion
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"